+++ apps/carp_mobile_sensing_app/lib/src/view_models/study_deployment_view_model.dart
@@ -1,4 +1,4 @@
-part of mobile_sensing_app;
+part of '../../main.dart';
 
 /// A view model for the [StudyDeploymentPage] view.
 class StudyDeploymentViewModel {
@@ -17,7 +17,7 @@ class StudyDeploymentViewModel {
   Stream<ExecutorState> get studyExecutorStateEvents =>
       bloc.sensing.controller!.executor.stateEvents;
 
-  /// Current state of the study executor (e.g., resumed, paused, ...)
+  /// Current state of the study executor (e.g., started, stopped, ...)
   ExecutorState get studyState => bloc.sensing.controller!.executor.state;
 
   /// Get all sensing events (i.e. all [Measurement] objects being collected).