+++ packages/carp_context_package/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.7.0
+
+* putting the location plugin back in businees (instead of the geolocator plugin)
+* revisiting issue [#389](https://github.com/cph-cachet/carp.sensing-flutter/issues/389), making a new implementation of the getCurrentLocation() method
+* improved the README to better explain how to handle location permissions on iOS
+
 ## 1.6.0
 
 * replacing the location plugin with the geolocator plugin